Front end Ember Invoice Generator completed

+++ app/assets/javascripts/controllers/fattura_controllers.js
@@ -11,18 +11,32 @@ App.FatturaController = Ember.ObjectController.extend({
             // set deleteMode back to false
             this.set('deleteMode', false);
         },
-        confirmDelete: function(){
-            // this will tell Ember-Data to delete the current fattura
+
+        confirmDelete: function() {
+          var transactions = this.get('model.transactions'),
+              list = transactions.toArray();
+          list.forEach(function(transaction) {
+            if (!transaction.get('isDeleted'))
+            {
+              transaction.deleteRecord();
+              transactions.removeObject();
+            }
+          });
+          var model = this.get('model');
+          if(!model.get('isDeleted'))
+          {
             this.get('model').deleteRecord();
-            this.get('model').save();
-            // and then go to the fatturas route
-            this.transitionToRoute('fatturas');
-            // set deleteMode back to false
-            this.set('deleteMode', false);
+          }
+          // and then go to the fatturas route
+          this.transitionToRoute('fatturas');
+          // set deleteMode back to false
+          this.set('deleteMode', false);
         },
+        
         edit: function(){
             this.transitionToRoute('fattura.edit');
         }
     }
   
 })
+
